+++ index.php
@@ -91,6 +91,7 @@ if(!$data_index) {
 $title         = $data_index->title;
 $subtitle = $data_index->subtitle;
 $summary = $data_index->summary;
+$featured_image = $data_index->featuredImage;
 $date          = $data_index->pubDate;
 $metak         = $data_index->meta;
 $metad         = $data_index->metad;